src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
huggingface_hub==0.33.0
pillow==11.2.1
transformers==4.52.4
torch==2.7.1
numpy==2.2
regex==2024.11.6
pybase64==1.4.1
pytest-shutil==1.8.1
diffusers == 0.34.0
